Refactored by splitting code among different files
